+++ src/migrations.ts
@@ -1,0 +1,13 @@
+import { PanelModel } from '@grafana/data';
+import { ClockOptions, ClockRefresh } from './types';
+
+export const clockMigrationHandler = (panel: PanelModel<ClockOptions>): Partial<ClockOptions> => {
+  const options: any = panel.options || {};
+  if (options.refreshSettings?.syncWithDashboard) {
+    options.refresh = ClockRefresh.dashboard;
+  } else if ((panel as any).refreshSettings?.syncWithDashboard) {
+    options.refresh = ClockRefresh.dashboard;
+  }
+
+  return options;
+};