+++ src/module.ts
@@ -3,5 +3,9 @@ import { PanelPlugin } from '@grafana/data';
 import { ClockPanel } from './ClockPanel';
 import { ClockOptions } from './types';
 import { optionsBuilder } from './options';
+import { clockMigrationHandler } from './migrations';
 
-export const plugin = new PanelPlugin<ClockOptions>(ClockPanel).setNoPadding().setPanelOptions(optionsBuilder);
+export const plugin = new PanelPlugin<ClockOptions>(ClockPanel)
+  .setNoPadding()
+  .setMigrationHandler(clockMigrationHandler)
+  .setPanelOptions(optionsBuilder);