+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
+## [1.2.0]
 
-## [1.1.1] - unreleased
+- Support local for date formats
+- Support refresh with dashboard time
+- Added dependency on Grafana 7.4+
+
+## [1.1.1]
 
 - Improved background